+++ mock/mock_node/mock_minimal_node.go
@@ -4,7 +4,7 @@ import "github.com/golang/mock/gomock"
 
 // NewMockMinimalNode returns a minimal implement of MockNode which accepts
 // any kinds of input and returns 'zero value' as all outputs.
-// Users can defines custom MockNode like this:
+// Users can define custom MockNode like this:
 // func NewMockCustomNode(ctrl *gomock.Controller) *MockNode {...}
 // in which users must specify output for all functions.
 func NewMockMinimalNode(ctrl *gomock.Controller) *MockNode {
@@ -18,6 +18,6 @@ func NewMockMinimalNode(ctrl *gomock.Controller) *MockNode {
 	mock.EXPECT().Stop().Return().AnyTimes()
 	mock.EXPECT().ReportExecuted(gomock.Any()).Return().AnyTimes()
 	mock.EXPECT().ReportStateUpdated(gomock.Any()).Return().AnyTimes()
-	mock.EXPECT().ReportReloadFinished(gomock.Any()).Return().AnyTimes()
+	mock.EXPECT().ReportStableCheckpointFinished(gomock.Any()).Return().AnyTimes()
 	return mock
 }